.gemini/antigravity/scratch/instagram-ai-news/requirements.txt
feedparser
Pillow
instagrapi
python-dotenv
requests
google-generativeai
beautifulsoup4
